Added stubs for the advanced instruction set (used for interrupts and calls).
Implemented "push" and "pop" because they are used by interrupts and calls.
Note - the code for interrupts is not tested (it is only an incomplete "stub").
Note - the default breakpoint and halt mechanics is implemented but the interrupt side is not.

Extended io.js with ready / hasData functions which will be used by the Teleprinter interrupt.
Because of this change "input" and "output" have a better way of setting the Overflow flag.